Add authorization context to App
create custom hook that uses context for user/loggedIn state. also
defines functions in context login, logout, and signup so any page or
component in the app can display dynamic components for whether or not
the user is logged in. Implementation begins with the NavigationLink
component which will display login/signup or chats/logout buttons
depending on context's isLoggedIn property