Verify logged in user via browser cookie
Defined useEffect in AuthProvider to, on component mount, verify that
the user on the browser has been successfully logged in within 7 days.
api-communicator will checkAuthStatus via a GET request to the backend.
The get request has middleware that will verify that the jwt is valid
and if so store the decrypted token id and email in a request cycle
scoped variable. verifyUser in the user-controllers will use this object
to find a user in the DB that matches the id parameter and send it back
to the frontend. In the AuthContext, the user state is then set to the
response object with the id and email and the isLoggedIn state is set to
true.